requirements.txt
# Core
flask>=3.0
flask-sqlalchemy>=3.1
flask-migrate>=4.0
pydantic>=2.0
pydantic-settings>=2.0

# Database
psycopg2-binary>=2.9
sqlalchemy>=2.0

# Async/Tasks
celery[redis]>=5.3
redis>=5.0

# Rule Engine
ciscoconfparse2>=0.7
jmespath>=1.0
textfsm>=1.1
lxml>=5.0

# Data Sources
python-gitlab>=4.0
netmiko>=4.0
ncclient>=0.6  # NETCONF
pysnmp>=6.0    # SNMP

# Utilities
python-dotenv>=1.0
pyyaml>=6.0
requests>=2.31
